fix: suppress ~550 code scanning alerts (batch 2) (#56)
Categories fixed:
- objectIndex (64): cppcheck-suppress in sha512.hpp
- cert-err33-c (159): (void) cast on printf/snprintf/fprintf return values
- cert-msc32-c (44): NOLINT on intentional constant-seed mt19937 in tests
- bugprone-implicit-widening (43): static_cast<size_t> on loop counters
- misc-const-correctness (37): add const to variables where safe
- shiftTooManyBitsSigned (16): suppress intentional sign-bit arithmetic shifts
- StackAddressEscape: NOLINTBEGIN in bench_field_26.cpp (benchmark escape idiom)
- nullPointerRedundantCheck (14): suppress defensive null checks
- arrayIndexOutOfBoundsCond (6): suppress bounded-loop false positives
- containerOutOfBounds (2): suppress guarded access false positive
- uninitMemberVar (2): init buf_ in RIPEMD160 constructors
- passedByValue (1): pass FieldElement by const ref (field.cpp/field.hpp)
- AssignmentAddressToInteger (1): fix pointer-to-integer in test_coins.cpp

Build: 48/48 targets OK (MSVC/Clang)
Tests: 25/25 passed

+++ audit/audit_ct.cpp
@@ -36,13 +36,13 @@ static const char* g_section = "";
     } \
 } while(0)
 
-static std::mt19937_64 rng(0xA0D17'C7C7A);
+static std::mt19937_64 rng(0xA0D17'C7C7A);  // NOLINT(cert-msc32-c,cert-msc51-cpp)
 
 static Scalar random_scalar() {
     std::array<uint8_t, 32> out{};
     for (int i = 0; i < 4; ++i) {
         uint64_t v = rng();
-        std::memcpy(out.data() + i * 8, &v, 8);
+        std::memcpy(out.data() + static_cast<std::size_t>(i) * 8, &v, 8);
     }
     for (;;) {
         auto s = Scalar::from_bytes(out);
@@ -55,7 +55,7 @@ static FieldElement random_fe() {
     std::array<uint8_t, 32> out{};
     for (int i = 0; i < 4; ++i) {
         uint64_t v = rng();
-        std::memcpy(out.data() + i * 8, &v, 8);
+        std::memcpy(out.data() + static_cast<std::size_t>(i) * 8, &v, 8);
     }
     return FieldElement::from_bytes(out);
 }
